fix: 예외처리에 대한 로직을 Validation 이라는 공통 클래스로 이동
